Split sys-start.r code into two functions
The comments in the commit explain it.  This is a step in the direction
of dividing the Rebol code used in startup between that which are needed
for basic interpreter function vs. doing things like printing out boot banners
or running startup scripts.  Further demarcating this division is a key goal
for Ren/C as a library.