Make stack overflow check macro more obvious
Being tired and bleary-eyed and wishing I could just get to committing
the stuff I wanted to commit in the first place, I completely misread and
mis-remembered what CHECK_STACK was and what it was for.

It's for *overflow checking* of the *C stack*.  :-/

Thinking it was something else I picked it up when trying to tidy up all
the Trap() variants.  So this makes what it is and what it does more
obvious, and also explains how it has no basis in the C standard and
probably another technique should be used.