Better way of handling errors in RL_Do_String
This routine is going away.  However the error outputting I added so that
using --do on the command line would print the error (instead of just
quitting silently) caused duplicate error prints in the interpreter.

It's a no-win situation with the RL_Do_String and RL_Print_TOS API
as written.  First of all, it doesn't print the top of the stack (it prints TOS1
which is bad)...and then it has no way to distinguish between an error
value that is returned and a caught error that is thrown.

This just goes ahead and prints the message if an error was caught,
and does not try and return a caught error on top of stack.  Instead it
puts an UNSET! there.