Improve molding of error object for THROWN()s
May be a longstanding problem where some branches would
crash on trying to form the VAL_ERR_OBJECT(value) of a THROWN
error.  I triggered it interactively in an older R3 download when I went
to look into a test with:

    >> try [catch/quit [attempt [quit]] print x]
    Segmentation fault: 11

This saves the fake created frame and uses that instead of trying
to get the one out of the error.