Fix halt in 'forever []' and similar (CC#2229)
Since I'm changing the error handling, I'm doing some tests of basics,
and in doing so I found out that you can't break out of 'forever []' in
Rebol. It's sort of metaphorical.

At the moment the interpreter will print out an error for the halt, but
that is an artifact of this particular state of the code...it's not in the
version I'm readying to commit.