Inline the one call to As_Port
The port-oriented code is able to notice if what it has is already a port
or not (e.g. a URL! or FILE!).  If it isn't then it tries to make a port out of
it.  That As_Port routine was doing a wily stack manipulation for what
could be done easier and more clearly.