Min_Max_Pair changed to not pass 'ds'
Passing the Rebol call stack frame as a parameter to a function when it
is not necessary just creates room for error.  It's better if all the D_ARG
and D_REF and D_OUT calls are made close to the entry of a native or
action and then pass the extracted values to the helper.