Fix for UNSET native returning uninitialized output
In Rebol's original formulation, the output for a function (D_RET, now
D_OUT in Ren/C) was initialized to an UNSET! before the call.  It
needed to do so because the value lived on the stack, and if the GC
ran it couldn't see uninitialized data.  But this meant a function could
think it was returning a meaningfully-set output by saying R_RET
(now R_OUT) to indicate it had written into the value.

This assumption is more likely to cause bugs than be of benefit, so
Ren/C is getting rid of it...and requiring a function to consciously
write to the output in order to return it.  The UNSET! native violates
the rule, and others may be found via the debug checks.