Get rid of Protected() in favor of GET_VAR_MUTABLE
Looking at the previous commit, I noticed the Protected() macro
for trapping if a word is protected.  It was only used twice with a
GET_VAR_MUTABLE already covering the check.  Between
that (if you want to trap) and TRY_GET_VAR_MUTABLE (if you
do not), protection status of a word is already taken care of.