Correctly chain T_Date() result value in REBTYPE(Time)
REBTYPE(Time) delegates to the date handler to do TIME + DATE and
reuses the code for DATE + TIME.  But as part of that delegation does not
honor the return value saying where the result of the operation is going.
(It assumes it's always R_OUT).  Which it apparently is for A_ADD as
written now, however one could imagine the implementation changing to
notice a zero time and just return the date as a R_ARG1, etc.