Series flag for IS_BLOCK_SERIES, eliminate SER_BARE
This attempts to kill several birds with one stone.  The motivator
was that IS_BLOCK_SERIES would test a series to see if it contained
REBVALs by seeing if its element width was 'sizeof(REBVAL)', as
opposed to having a series flag for that purpose.  Series are used for
many other data types (pointers, REBYTE, REBUNI, etc) it always struck
me as sketchy because you might want to use it to make a series of
REBVAL-sized things that aren't REBVALs, and indeed the STRUCT!
implementation wound up doing that (in 32-bit builds).

This called for a review of the somewhat-related SER_BARE flag, which
would prevent the garbage collector from descending into a series for
marking based on a guarantee that the series contained no composite
values.  This wasn't to say that it didn't contain REBVALs, in fact the
only original instance of these were for lists of REBWRS values (a.k.a.
"Unwords") in function word series and object frame keys.

The "take my word for it" property of SER_BARE was not as good as
having a debug check, and the "bareness" was already contextually
known to the GC based on the type it was marking.  This replaces that
flag with SER_BLOCK and adds asserts regarding the "bareness" into
the appropriate cases in the Mark_Value switch statement.

This also changes MARK_SERIES_SHALLOW to the better cueing name
of MARK_SERIES_ONLY, and uses it for ROUTINE_FFI_ARG_TYPES
as @ShixinZeng pointed out that these are not block series.