Fix CRC32 sign regression, reorg/explain CHECKSUM
In trying to stomp out sign comparison warnings, I'd changed a loop variable
'i' from a signed REBINT to an unsigned REBCNT in CHECKSUM.  But the
variable was reused for and serving as a static cast for an unsigned CRC32
to a signed value.

Tests caught the change, so I fixed it.  While looking, I noticed that the
Adler32 checksum was returning an unsigned value, which wasn't obvious
from looking (and in fact you can't jump-to-definition for 'z_adler32', because it
is one of those macro-renaming tricks).

So to make it clearer what was going on, I gave each one of the checksum
methods its own variable to type appropriately.  That way it could serve as
documentation of what is actually being returned in each case.  Then I left the
loop index unsigned to avoid the comparison warning.