Fix tuple overflow regression
In another case of a reused variable that has one application in signed
vs. another in unsigned, the 'v' variable was used to loop through tuple
indices as well as to hold a signed value (which may overflow).  Having
found two of these cases now, I'm going to look around for any other
REBINT/REBCNT switches that may be having effects not caught in
the tests.