Fix regression on HTTP reads from ea26f4a
In switching the Function-based DO routines to not return a pointer
to a popped stack element, I'd somehow neglected to update the
parameters in wake_up to reflect that they should be looking in
DS_TOP instead of in the reused pointer 'val' that had previously
been returned.

(It also doesn't balance the stack, but Rebol doesn't require that...
yet.  Soon.)

Thanks to @draegtun for building and keeping up to report it.