Make comments and names consistent for GC helpers
This enforces a uniform convention on how the GC routines are named
with regards to terms "Queue" and "Deep", and also makes sure that no
calls to deep are made on non-blocks.  Helps enforce the use of the
proper macros by having the macros set the mark prior to calling the
core routine, so if someone tries to use the core they'll get an assert
about the series not being already marked.

Also adds assertions about the mark stack being empty on entry and
exit to the GC.