Fix unicode path bug on single-byte char systems
One of the routines I'd switched to use a "p" instead of a "bp" I apparently
did not update the references to use the p.  :-/

Which would be a fairly lame oversight, but something strange must have
happened because I was pretty certain I'd gotten rid of the GET_ANY_CHAR
macro... but here it was still in existence somehow.  The likely issue is that
I screwed up a merge somewhere.

Resolves subdirectory iteration issue reported by @Maarten on SO chat.