Make REBDOF take a const pointer to function-valued type
This change helps reduce the likelihood of an accident when passing the
new REBVAL* argument to the DO routines, in distinguishing it from the
REBVAL* holding the function by being mutable.

It was a little surprising that ROUTINE! when dispatched modifies a series
kept inside the REBVAL* for the routine on each call.  I've reported this to
@ShixinZeng as a problem for threading (if and when that ever happens)
so there's a mutability cast there for now to keep track of the issue.