Do_Next and related routines allow arbitrary output location
This is a graft of the interface of StableStack which still uses the Data Stack,
so that you always supply Do_Next with a place where it should put the
result of the evaluation.  This is more efficient, easier to read, and easire to
keep the stack balanced.  It is still possible to write to top of data stack as:

    // preload output with a value that in debug builds is considered valid
    // for intermediate GC, but invalid for all other usages
    DS_PUSH_TRASH_SAFE;
    Do_Next(DS_TOP, block, index, FALSE /* not an op */);

The intermediate solution is to make the output slot hold a REB_HANDLE
that is a pointer to the actual location where you want to write.  Because
the GC skips over handles and doesn't mark them, it doesn't know the
handle contains a pointer to a REBVAL...so there is a special scan of the
stack frames with Mark_Call_Stack_Frames() to keep anything written into
the output location safe from GC.

This also changes it so that when the stack is empty, the DSP is -1 and not
zero.  That's because the stack is indexed directly via the zero-based DSP
value, so an empty stack can't be indicated by a DSP of 0.