Balance stack for ordinal natives and LAST
Although FIRST/SECOND/etc. and LAST are single arity, they are
generalized and use the action routine for PICK...which is arity 2 in
the action dispatcher.  These troublesome natives thus "fake" their
call stacks by pushing an integer onto the data stack and assuming
that will make it the D_ARG(2).

It's an assumption that won't be true soon.  But for now this balances
the fact that they didn't pop the integer after pushing it, so that the
new check in Do_Next is happy.