Core Do implementation can process to block end
This introduces a flag to Do_Next (now called "Do_Core") so that it has
a condition to keep evaluating until there's nothing left to process.  Then
there are two convenience macros: DO_NEXT and DO_BLOCK, which
are exposed to invoke these behaviors (while not having to pass in the
OP flag that is used internally to the Do processor).

It also adds a THROWN_FLAG which can be tested as the result of a
call to the core Do function.  This is more efficient than having to test a
result for both being an error and being a "throw-class" error, but also
lets DO_BLOCK easily return either a "yes I made it to the end without
a throw happening" or "no, a throw happened and I didn't reach the end"
result.  That means that constructions which used to be two steps as
DO_BLK followed by an IF (THROWN()) are just IF (!DO_BLOCK())

Then DO_NEXT has three result states: END_FLAG, THROWN_FLAG,
or the index of the next evaluation.  If you have a block like [2 + 4] then
the first call will give you back an integer 6 and an index 3 (zero-based)
and the second call witll return END_FLAG and your value will be UNSET!