Assertion failure and printout for "early errors"
If some kind of error happens during an init, it's more convenient in the
debug build to assert than to throw it and lose the stack.  This precludes
the use of TRY and ATTEMPT in the early boot, so perhaps it will have
to be softened at some point if that turns out to be essential.  For now,
it isn't...and this is helpful.