Get rid of DS_ARG and DS_REF
These routines are implicitly dependent on the DSF in effect, while
D_ARG and D_REF are explicitly dependent on the call frame passed
to an action or native.  The explicit dependency is preferable in cases
where it is available... and if it's not available, then using DSF_ARG
and DSF_REF is preferable.