Some mutability casts needed for the C++ build
I mention in the comments for VAL_EVENT_SER what the problem is here;
it's related to the casting trick the macro uses, and that trick is needed
because the series is not defined as a REBSER* in the API.  m_cast is
ugly, and in some of these cases not harmless (const data is actually
being updated).  But these uses at least it keeps track of such things and
makes them glaringly obvious.