Simplify misguided "optimization" of REBUNI strlen
The general idea was sound (use possibly vector-optimized memory
operations instead of a byte search), but when memmem() was not
available in POSIX I fell back to a strchr() search for zero, and then
on each call checked if the hit were on an even byte boundary.

Well and good in theory, and if you don't use the Latin alphabet then
probably a good idea.  Pathologically terrible in practice, as almost
all the REBUNI characters in wide strings have a leading zero byte
when processing code written in mostly-English.

memmem() for a two byte zero sequence with a similar premise
to the previous code could work, if that optimization is interesting
enough to put in a conditional behavior for GNU.  Not a priority at
this time to worry about it.

(Lesson for the day: even if you know your code "works", step
through it anyway.)