Fix reused stack slot memory for Do_Args() result
When a thrown error is reported by Do_Args, it leaves it on the top
of the stack. (Note: This has been completely rewritten in the
StableStack implementation.) The expectation is that the caller
will look at the top of stack after a call to see if it was thrown.

By using DS_SKIP inside of Do_Args, it merely advanced the DSP
over what would become the top slot. Hence if a function had all
its arguments quoted and ran no evaluations into DS_TOP (e.g.
HELP) it could leak whatever previous information had been on
the stack as what the caller saw. If that was a THROWN() class
error (a RETURN, THROW, BREAK, etc) then it would appear a
Do_Args had failed when it had not.