Reuse work from previous OP fetch in DO
Formerly, the Do_Next function would evaluate one expression at at a
time.  There were "do an entire block" drivers outside of it to do more
than that, but now that doing an entire block is an option inside of the
core evaluator loop itself...it's possible to reuse the result of the fetch
of an OP! for a WORD! that occurs at the end of every evaluation (
when an evaluation is followed by a word).