"Chunked" call stack implementation
The comments in sys-stack.h and m-stack.c summarize this.  But the
idea is that with call frames needing to be stable w.r.t. where the pointer
to their arguments are while a native is running, then you want
something more clever than an Alloc_Mem() and Free_Mem() for the
call frame every function call.  This lets you make "chunks" of allocated
memory where the call frames are put, and uses a heuristic of always
keeping one extra chunk available for overflow to avoid memory churn.

With this modification, all the changes in StableStack seem to give a
performance profile roughly equivalent to the rebolsource build in
casual tests.