Get rid of MUNGWALL-related defines and calls
There was apparently a memory checking utility on the Amiga called
"mungwall".  The internal-to-Rebol bounds checking feature was thus
named after it, although it implemented a "mungwall-like" functionality.

The use of Valgrind and Address sanitizer are a less invasive form of
verification, and more powerful as well.  In addition, we are using the
ASAN API to manually poison and unpoison memory.  The MUNGWALL
code is really just in the way--and especially so as no one compiles
using it.