Catch mismatched allocator use in debug build
Historically it's been a little hard to understand when to use the various
allocation and free routines and exactly what the differences are.
There are now some comments and structure surrounding malloc() vs.
Alloc_Mem(), vs OS_Alloc_Mem()...but under the hood they were all
doing about the same thing.  (Alloc_Mem doing calloc() to zero fill,
but that will change.)

To help catch mismatches in usage, this changes the implementation
so that you can only use a free() with its associated alloc().  It also
adds a debug check to ensure that calls to Free_Mem() are passing
a correct value to match the size of the Alloc_Mem().  (Knowing how
much memory being used is important to the GC, so it should be right,
although any allocations that are done with OS_Alloc() or direct
malloc() will not be counted in that amount.)