Get rid of SER_FREE flag
The SER_FREE flag was not used and was just confusing.  The real
mark of a series in the pool being freed is if its "width" (element size) is
zero, which is actually a better way of doing it because it means most
series operations simply will not work (vs. setting a flag that few
operations would bother to check, and shouldn't be paying for checking
as it's an ideally impossible situation to get references to free series)