Unify function dispatch under Dispatch_Call
This brings together all the places that invoke a prepared call frame
to go through a common function for dispatch, so that they don't each
need to know how to dispatch the different types...or to do the push
and call frame freeing.  It also means they'll go through a common
path for things like return type checking and debug checks for
stack balancing.