simplify value options, reclaim OPTS bits, rename unword
The header for every Rebol value is limited to 32 bits of functionality
(for common behavior on both 32 and 64-bit platforms).  These 32 bits
are chopped up into 8 for the datatype, 8 for options that would apply
to a value of any type (e.g. the 'newline' option that can apply to any
value in a series), 8 more bits for options that are type-specific, and 8
bits reserved for future use.

It so happened that some options that only apply to words were living
in the options that could apply to any value.  I moved these into the
word-specific bits, and while I was at it changed the poor name "unword"
to "typed word"--since it contains a typeset instead of a binding.

I also flattened out the macros to make it more obvious what is going
on.  If you have a macro for SET_FOO(value), and all it really is doing
is SET_FLAG(value, FOO_FLAG)...then it's more confusing than
anything.  (Especially bad when you have macros named things like
MARK_SERIES which all they do is set a flag, because the implication
may be that they do more work than that, when they are very simple.)