Add missing VAL_INDEX assign in Scan_Any
The Scan_Any routine bills itself as being for scanning a string that
"doesn't require any special decoding"...which today means tags (as
they don't currently have escaping).  Be that right or wrong, there's an
apparently long-standing issue where the index was not being assigned
in the created value.  Artifacts of implementation covered that up, but I
moved some things around and uncovered it.