Catch use of VAL_SERIES on non-SERIES! types
The Reb_Series_Ref type and the Reb_Object both had a REBSER* in
the first slot of the structure; as the series being referred to or the frame
of the object respectively.  This meant that if you used the VAL_SERIES
macro or VAL_BLK_* macros they would work on either, despite the
overlap being incidental.

Static-typing is the best way to catch this, and there's a solution for that,
but these were some basic violators caught with a simple runtime test.
Since the test relied on making sure VAL_SERIES and VAL_INDEX were
only used on a valid Reb_Series_Ref, it meant some initializations had
to be reordered to set the type first and then the series.  These changes
are included as well.