disambiguate VAL_DECI from VAL_DECIMAL
Going through sys-value.h and trying to document what I could,
the annoyance of how a datatype called "deci" is used for money
when there is also a decimal type crossed my radar (as it has
before).  To fit it into my idea for formatting everything I went
ahead and made the money type more clear.

I went with VAL_MONEY_AMOUNT instead of just
VAL_MONEY, as it may be that money gets back its feature of
having a currency code in it and being some kind of composite
that's more than just the number.  Because I don't know what
to call the 'deci' type I left it as is for now.