Normalize and document Reb_Value and its parts
This is a documentation sweep of Reb_Value, incorporating what
I found out over time about some of the hidden connections
between value types.  It's also a general cleanup and adds some
comments.

It's also an attempt to reclaim some names for things that were
not used often in the code.  e.g. REBWRD was not invoked often
(or at all) when it referred only to 3/4 of a REB_WORD value.
Yet it could be much more useful as a type for documenting when
a routine expects to be passed an ANY-WORD! value... (and
even more useful still if it could be used for static type checking
of that fact.)