[catch] function spec block exemption
Some common routines that are Rebol2/Rebol3 agnostic use the
"catch" attribute from Rebol2:

    func [[catch] ... ] [ ... ]

This doesn't do anything in Rebol3, and has just been ignored.  The
<infix> change marked the shift toward using tags for attributes on
functions, so this was removed...but since there is still code out
there for now, this one exemption has been put back.  This prevents
the skipping of arbitrary block data that occurs before parameters,
and documents what's going on so people reading the code know
what's up.