GC obsolete parse code, keep dead /ALL refinement
Rebol2's notion of PARSE was to do "magic" to let you not have to
worry about whitespace string parsing, and you would only get all
string characters if you said /ALL.  In practice, the "magic" wasn't
good or feasible to be made good, so people used /ALL.  Rebol3
turned on /ALL by default for block rules.

Which leads to the next issue: PARSE had some other modes
besides passing in a block.  Given how deep and complex PARSE
is already, it doesn't make sense to overload it with more options,
and improved alternatives like SPLIT exist to choose from for the
functionality these options represented.

This marks /ALL for death but leaves it still for this moment to keep
common subroutines with Rebol2 from erroring when it is used.