EXIT acts as QUIT if not in function, LEGACY flags
This adds a new behavior for EXIT (as it tries to find a place and
meaning in a definitionally-scoped return world).  By virtue of not
having a binding and yet still wanting to exit "wherever it is", the
semantic fuzziness of EXIT is potentially served by meaning "exit
from wherever it is you find yourself", so rather than having EXIT
say "no function to return" from when invoked from the shell, it just
acts like a QUIT.

Moving forward on this distinction also lets EXIT mean "exit from
the currently running parse paren", which is also not a "definitional
scope".  To be more specific about where you want to go, you can
use a definitional return or a THROW/CATCH.  These features will
be coming shortly.

But for the moment, this just twists how EXIT works w.r.t. the top
level, so you can say EXIT to get out of the console.  It also adds
the first legacy switch--available in the debug build--if you want to
make it so that exit will say "return not in function" instead.  To get
that, set an environment variable R3_LEGACY=1 and then when
running say:

     system/options/exit-function-only: true

This is a somewhat mundane first test of the idea, with a behavior
unlikely to cause too much trouble  either way, to see how it works.