Gabriele Santilli's MATH and FACTORIAL (CC#2120)
This lets you do math evaluation with "usual" precedence rules:

    >> math [1 + 2 * 3]
    == 7

It also has a mode where it will merely do the transformations, which
had been called /TRANSFORM but I called /ONLY for the sake of
its representation of a simpler step (e.g. 'do math/only x' is 'math x')

    >> math/only [1 + 2 * 3]
    == [add 1 multiply 2 3]

Whether this is an ideal implementation or not isn't really the point.
As CC#2120 argues, having it to demonstrate shows that Rebol's
idea is to make the engine simple and predictable, so you can create
your own if you need to.  (And you usually stop wanting to.)