All functions ending in ? return LOGIC! (part 1)
Though it looks simple, there was a bit more to this than just a
search and replace.  Still it is about a list of just five transtions:

LENGTH? => LENGTH
TYPE? => TYPE-OF
INDEX => INDEX-OF
OFFSET? => OFFSET-OF
SIGN? => SIGN-OF

Though they are renamed, there are still legacy definitions.  So
the sky isn't falling for code written using LENGTH? or INDEX?.
(Yet.)

This doesn't actually take care of *all* the ? routines...just the
common ones which consensus and conversation have focused
on.  Better names and study are still needed for bind?/bound?,
encoding?, file-type?, speed?, suffix?, why?, info?, and exists?.

A Porting Guide has also been started:

https://trello.com/b/l385BE7a/ren-c-porting-guide