Fix THROWN() handling for PAREN! in PATH! (CC#2243)
This is not a complete solution to the problem of "things being thrown
during a paren path evaluation".  A real solution for all the callsites
will come when DO_PATH returns TRUE or FALSE the way that
DO_BLOCK does, and when a single output value is enough to hold
the result of a path evaluation (via whatever implements REFINED!)
But this moves the problem to callers not paying attention to the result
instead of Do_Path() not providing it at all.