Make CONTEXT native pass on THROWN()s
I was going over pending issues with warnings related to not paying
attention to the result of DO_NEXT.  And I noticed that the context
creating native didn't handle throws.

That seems like it would have been a problem, and it would have...
if people were *using* the context creating native.  However, this
was only used as a buggy substitute for "make object!" in sys-obj.r...
later overwritten by a user function in the mezzanine that called
MAKE OBJECT! as its implementation.

I'm not a big fan of CONTEXT as a verb form for MAKE OBJECT!
nor OBJECT, and I prefer re-purposing HAS (aligns with DOES).
But still, it seemed to make sense to fix up the native for now and
delete the mezzanine, just to make things more clear.