New CASE with switch for legacy behavior (CC#2245)
This reworks and elaborates on the implementation of CASE (with
comments).  Because it winds up being a fairly radical change, it
has a switch called broken-case-semantics that can be turned on
in debug builds when R3_LEGACY is set in the environment,
which should give it the behavior of an old Rebol2/Rebol3 switch.

A main change is to address CC#2245, where 'case [true add 1 2]'
and 'case [false add 1 2]' have differing interpretations of structure
of the case.  In Rebol2 they return 3 and 2 respectively, while under
this implementation they return 3 and none.

A second change is the removal of a "feature" which was to tolerate
an odd number of condition/body pairings.  In Rebol2/Rebol3 this
would run the condition and then return TRUE or NONE.  Thus
'case [1 > 2]' would be NONE and 'case [1 < 2]' would be TRUE.
While on the surface this seems to be just actively asking for bugs
to be harder to find, it was found used once as a way of being
lazy and not putting a "true" for the condition slot of a last default
item.  But TRUE as condition of the default body in the case is a
fine idiom, is cool, and doesn't open the door as wide to bugs.)

Third is that the return value of the case is always the evaluated
value of the last TRUE? case that was run.  This means that:

    >> case/all [true [1 + 2] false [9]]
    == 3 ;-- would have been none in Rebol2 and old Rebol3