Fix leftover warnings of unused result for DO_BLOCK
When a DO_BLOCK operation inside of the C code ran historically,
(then called DO_BLK), it would execute the block to its end and then
return the result on the top of the stack (technically in the former top
of stack position that had been popped, hence unstable, but that has
been ixed.)

By returning a result on top of stack implicitly, it was not always the
case that this was checked to see if it was THROWN() e.g. some
THROW or RETURN or BREAK kind of value (that it should bubble
up the stack if it didn't itself process).  Several of these were bugs,
but some implicitly "just worked" because they were at the end of
an evaluation and were going to return the value anyway.

Shifting DO_BLOCK to return a result of TRUE for the block being
run to completion and FALSE if there was a throw saves on needing
to test the value for both its type being error and that it is a "thrown"
class of error.  It also means that -Wunused-result can serve as a
trigger for catching anyone who doesn't save the result or have
conditional behavior on it.

This catches the stragglers who didn't represent bugs by explicitly
ignoring the thrown status (and explaining why they can).  It also
locks down the system calls which "know" they can ignore the
value result of some Rebol code by ensuring that there isn't any
result (by way of requiring it to explicitly be UNSET!)