Quarantine PARSE's hacky use of longjmp()
There was a lot of terminology and implementation confusion
surrounding how Rebol used the concept of "throwing", and what
it meant to be a "thrown error" or to have a value test positively for
being THROWN().  Some uses would longjmp(), while others would
return and bubble up on the stack.  Some returned arguments and
others could not.  PARSE introduced the use of something it called
a "virtual return" and it was historically murky:

https://github.com/rebol/rebol/blob/25033f897b2bd466068d7663563cd3ff64740b94/src/core/u-parse.c#L1267

Over time it has emerged that only two cases of what would be
called a "throw" (that intended to convey a payload up the call stack)
ever used a longjmp vs. just bubble up.  This makes one of those
two cases clear in PARSE.  While it should really return up the stack
in a normal way, this at least takes a step toward it anddocuments
what it's doing much more apparently.  It also paves the way for
switching "Traps" to always take just an error code, and never a
REBVAL as an argument.