Do not try to "bring-to-front" console window
With a GUI-mode binary that creates its own console on demand, this
was probably sensible, so that the freshly allocated console window was
not hidden behind a bunch of other windows. But for a console-mode
binary, the console is managed by the kernel, with the corresponding
window also managed accordingly; so for the console-mode binaries we
currently support this doesn't make much sense.