Don't set a custom console title
This was originally mainly used so that the console window could then be
found with FindWindow so that it could then be brought to front. But as
a plain console-mode binary, we no longer do that anyway. Leaving the
console title unchanged makes it easier to discern different binaries
when running multiple, differing Rebol 3 binaries in parallel.