QUIT via THROW, CATCH [] vs CATCH/ANY (CC#2247)
(Note: Should fix #41)

In the commit description for 2db2935d92 I explained how THROW
had two cases where the longjmp mechanic was used to pass a
value up the stack, instead of "bubbling up" through stacks that did
not have to explicitly initiate a setjmp state or recover from a longjmp
in order to receive the values.  The first was when PARSE needed to
"prematurely" end the parse process (through something other than a
match failure).  The second was when a QUIT was issued.

This takes away the requirement for CATCH to do a setjmp setup by
making QUIT a named throw.  Propagating that decision through
led to finding some more "swallowing" cases where THROWN values
were being ignored and not handled in any way, such as if there were
a THROW out of a MAKE of a PORT! or MODULE!.

There is a behavioral change to CATCH needed for this, which is that
it does not catch a named throw if you do not use /NAME.  Without it
this can't work.  Because some clients (notably test code) need a way
of actually intercepting any THROW, an /ALL refinement was added.
Expanding the range of values that can be used as a /NAME may
change what is used for QUIT to not be a WORD!.