Inline Win32 console initialisation code (Init_Console)
As we are built as a console-mode binary, the console is ensured to be
present by the kernel. Hence we can save the AllocConsole call [*] and
getting the stdio handles twice. At which point, not much of
Init_Console remains and inlining it again improves readability.

[*] There are cases when a console-mode application still doesn't have a
console, such as services which run their executable as a detached
process. However, for these cases it's also better to _not_ override the
intention of the caller. So we better do not try to explicitly allocated
a console but rather assume that if we don't have one, then that's how
it's supposed to be.