!DO_BLOCK(...) => DO_BLOCK_THROWS(...)
Given the now fully-apparent pattern in code that you always need
to test a DO_BLOCK to see if the returned result is thrown, I found
that it would be more readable to invert the condition to show exactly
what it was being tested for.  This removes the need to explain at
every IF statement that the false branch is because a value was
thrown.  I use the present tense "_THROWS" to help cue the
reader into realizing it's speaking in the "present sense" of the block
being evaluated that you are passing.