Expand possible /NAME types for THROW
This is a change which unhinges the THROW and CATCH mechanics
from the implementation details of ERROR!. Keeping them separate
is good for continuing the stamping out of confusion which these
types have experienced, which helps make sure that throws don't get
dropped on the floor. But additionally the switch of THROWN() to an
options bit (that can be applied to any value) permits the /NAME of a
throw to be derived from the full spectrum of the rest of the value.

The prior limitation of name specification to a 32-bit slot inside of an
ERROR! value meant that a named throw could only have the
interpretation of that as a word symbol ID. So only unbound words
could be used as /NAMEs. This change will *mechanically* permit
any name, but currently it limits to NONE!, WORD!, ANY-FUNCTION!
and OBJECT!. Example:

    >> catch/name [throw/name 1020 :append] :append
    == 1020

For control-flow mechanics like BREAK, CONTINUE, and QUIT this
currently uses WORD!, but can be tuned to other choices as they
are chosen to be fit.  It also does the equality test via EQUAL? and
hence it will ignore the binding of words in a CATCH/NAME, which
doing differently may offer technical advantages.  Either way it gives
the basis for whatever the best solution is.

This commit also adds /WITH as the preferred way to specify an
argument to BREAK.