Pickups for DO_NEXTs not handling THROWN_FLAG
A quick audit of DO_NEXTs that didn't test THROWN_FLAG
(or test it after looking at the value, which is now a bad thing,
as the value would be a throw label and may be a type you
are interested in as opposed to an error which callers
probably were not paying attention to).

Additionally, I changed the name to DO_NEXT_MAY_THROW
as an attempt to alert those using it to be attentive to the
existence of THROWN_FLAG.  In general the convention of
moving everything into the interface of the function instead of
implicit on the stack or some property of the stack value has
cleared things up a lot.