Skip stack-relative lookups in forming call frames
I knew it was possible to leak out a word from a stack relative
frame lookup and then try to look it up in a forming frame, but I
figured any code that did so would be broken and so an assert
would be good to notify people of that.  What I missed was that
the mechanics of the previous code was such that forming
frames...though positionally deeper in the stack than the frames
which would sneakily look into them...were still not considered
as candidates for stack-relative lookup.

Which is a long way of saying Rebol3 already gave an error on
cases like this, when I thought it was letting it slip:

    leaker: func [/eval e /gimme g] [
        either gimme [return [g]] [reduce e]
    ]

    leaker/eval reduce leaker/gimme 10

So my assert really should have been an error, and now it is.
I'll note that it's picks up an improved error, as before it was:

** Script error: g word is not bound to a context

But now it is:

** Script error: g word is bound relative to context not on stack