/WITH for CATCH w/handler, TRY => TRAP (+ legacy)
This adds the missing ability of CATCH to supply a handler, via the
/WITH refinement.  It is modeled after how TRY worked, which was
to take a BLOCK! or an ANY-FUNCTION! and be willing to take a
function that can process its args -or- one of reduced arity.  The arg
order in CATCH's case is thrown arg first, and then the throw name.

It also renames the TRY native to TRAP, because it's basically just
better in every way...but leaves TRY around for compatibility.  What
makes it better is that it doesn't have the TRY/CATCH baggage
where people will think you use it to catch throws, but also that as
a word it's less ambiguous on the fact that it evaluates to the error.
("TRAP" is more parallel to "CATCH" in that it's an action where
you expect to have a thing in your hand that was caught/trapped.)
TRY may be a better and shorter fit for ATTEMPT where you try
something and then if it didn't work out you just have NONE!

While I was at it I made it so that giving a function expecting more
args than the handler could provide would be an error (accounting
for refinements as not being required arguments).  I also added a
more explicit message about trap needing you to pass it a function
that takes an ERROR!, but mostly just because there was a
comment saying it might need a better message.  (As I made that
note I could have just deleted it, but I figured it wouldn't hurt to
have a more specific error.)