Use more compatible dir testing than DT_DIR
*Some* filesystems on *some* operating systems support
the ability to tell you whether a file record is a directory or a
plain file with a flag that you get during the enumeration of
file names in a directory.  Unfortunately, because it is a
feature that is filesystem-specific...just because your OS has
support for it doesn't mean it will work.

Case in point: I found out it does not work with VirtualBox's
shared folder filesystem (which I've switched to using instead
of the painful NFS method I used previously).  This apparently
is an issue with XFS as well.

Fortuitously the code for the more widely compatible method
is already in there, as it was needed for HaikuOS.  When the
day comes that the speed of Rebol's ability to tell if something
is a directory or not while walking the filesystem becomes the
linchpin of its success, we can add feature-testing heuristics
to see if the filesystem actually supports the flag or not.