axe commented-out items: as-binary, as-string, alias
Alias was already gone as a REBNATIVE, though it had an error
message handing around.  The bodies for as-binary and as-string
just gave a "deprecated" error (that you couldn't get since there
was no definition for them).

For the historically curious: as-binary and as-string were holdouts
from when strings had one byte per character so it was technically
possible to get a "view" of them as either type, and modify such
that other views of either type would see the changes.  As chars
no longer map one-to-one to bytes, this isn't technically possible
(and may not have been a good idea to start with anyway, although
we do face that question of "views as different type" with other
series types where it is still technically possible).

ALIAS let you give two different words the same symbol ID, which
was just deemed a bad idea all around.