Force SERIES_GET_FLAG to boolean-safe result
The macro for GET_SERIES_FLAG used bitwise logic to get
a result that could be tested for truthiness by an IF statement.
But because it was missing the != 0 test you could wind up with
the bit set in a high location, which would be a problem in
an assignment like REBYTE cond = GET_SERIES_FLAG(...)

Changed flag clearing and setting routines to cast to void to
avoid any leakage problems.

O...the joys of debugging low-level C code.  :-/