REBYTE wide of series, clarify is_wide char vs. "wide"
In looking at the interface for creating a series, it took two numeric
parameters "length" and "wide", the latter being the element size
(which is constrained to a maximum of 255).  As it's easy to get
them mixed up, I thought it would be good to change the "wide"
to be a REBYTE to help reveal mixups via type check warnings.

(This doesn't have the interface change--separate commit.)

Search pointed out an issue which is that "wide character" as a
unicode term is used here and there, so "wide" can wind up being
a boolean quantity.  This suggests avoiding the term when
possible in the Rebol codebase and sticking to more formal terms
like UCS-2, UTF-32, etc.

(I'll also note that the word "wide" is used to describe the size of
series pools in a case where the pool itself is not limited to just
a byte in size.)