Fix sketchy CASE defaulting usage in ARRAY
Array was making use of something you were previously allowed
to write:

    case [
        condition1 [body1]
        condition2 [body2]
        condition3
     ]

Support for this was eliminated in cc08fea in an attempt to make
CASE better defined.  It's hard looking at the above to know what it
might return.  If one is to believe that allowing uneven numbers of
condition + body isn't just going to make bugs more confusing, one
might think it would be an okay way to provide a default action or
value.  However it does not do that...the return result is TRUE if
the condition matches and the result of condition's evaluation is
thrown away.

All things considered it's better and more regular to just use TRUE
for the condition and then put what code you like as the body.
That permits even running things that don't return a result (which
would cause an error for the condition slot).