Uniform function-based Value Initialization
One trick to make the managed/unmanaged distinction less invasive
(upcoming commit) is to make it so that there is an implied delegation
of memory management to the garbage collector for any series that
is put into a REBVAL.  That lessens the burden on routines to
manually MANAGE_SERIES.

Some of the value initialization methods were already functions,
while others were macros.  As long as I was going around changing
them I went ahead and renamed them...as routines like SET_MAP
were not very clear.  It sounds more like an operation for putting
things into a map.  And the SET word is rather overloaded as it is
(variable setting, set words, and sets on the contained thing) that I
felt that Val_Init was a better choice.

This kicks off a convention of allowing Caps_With_Underscores
naming conventions for macros that use their arguments exactly
once.  That marks them as being "as safe as a function" w.r.t.
argument evaluation, and is a contract to not change that behavior.