Simple GC disablement during Debug Output
Pinning down the exact moments when a GC can happen needs
to be done.  For now, one place it can happen is during I/O and
that includes printing.  Debug routines like Debug_Fmt() and
PROBE use the same I/O as everything else, hence they can
trigger the GC.  This can be bad news if you're in the middle of
code that assumes no GC is possible.

This is a quick disabling of GC, using the GC_Disabled flag and
not ENABLE_GC/DISABLE_GC (as those are going away in a
subsequent commit, and generally not going to be a choice for
non-Debug code to do).