OS_CALL_PROCESS uses OS_ALLOC, not malloc
When I was initially doing the Coherence 1.5 work, I didn't want to
get too invasive on the unfamiliar CALL code, considering everything
else.  However I did notice that it was using malloc() and not
OS_ALLOC to create buffers that it would hand back to core, which
meant the core would have to use free().  Then I noticed it was
using realloc(), and there was no OS_REALLOC hook point (and
shouldn't be one).

So I punted it and left an alert comment about it.  But then later when
I changed it so that the OS_FREE and FREE routines would pick up
on mistakes with the allocator, it caught it.  And it was trivial to just go
and memcpy the buffer into a larger buffer.