Free temp buffers in Modify_Series, clarify "Append_Byte"
This changes Modify_Series so that if it creates a series as a
temporary in order to perform the operation, it frees that temp
rather than leave it for the GC to pick up.

In trying to formalize the responsibility of memory management
and who is responsible for freeing what, I ran across the fact that
the routine Append_Byte was doing double duty as a way of
making a new series if the target of the append was NULL.  As a
surprising behavior that required comments at the callsites saying
what was going on, it seemed better that function be broken out
and named something like Make_Series_Codepoint.  (One might
consider Make_Codepoint_Series, but that sounds a bit like you
are making a series of multiple codepoints...so rather I went for
more like Make_Series_From_Codepoint, just not as long.)

While looking I noticed something else, which is that Append_Byte
doesn't append bytes either.  It will pick the size of what it appends
based on the target, and not expand the target width to fit it.  For
something that strange it needed a stranger name, so I went with
Append_Codepoint_Raw.  Which at least will get you to look at the
definition instead of making assumptions about its behavior.